Render account list as template
The account list currently has the prefix/suffix interpolated into the
template, which makes it hard to see what the actual email addresses are
in production account lists.

This changes the scripts to render the account list as a template,
allowing some values to be hard-coded into the final manifest so that
you end up with more readable email addresses.